+++ gen_docs.sh
@@ -3,4 +3,5 @@
 # to generate docs, first run:
 # sudo -EH npm install -g jsdoc jsdoc-to-markdown
 
-jsdoc2md ../src/context.js >./context.md
+jsdoc2md ./src/context.js >./docs/context.md
+jsdoc -d ./docs ./src/context.js